% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast}
\alias{cast}
\title{Cast a long dataframe into a wide dataframe}
\usage{
cast(
  df,
  row.var,
  col.var,
  value.var,
  fill.value = FALSE,
  strings.as.factors = T,
  verbose = F
)
}
\arguments{
\item{df}{A long form dataframe to cast}

\item{row.var}{Column name for the values to map to the rows of the output dataframe}

\item{col.var}{Column name for the values to map to the columns of the output dataframe}

\item{value.var}{Column name for the values to map each cell in the output dataframe}

\item{fill.value}{A value to fill cells. Default: `FALSE`}

\item{strings.as.factors}{Convert strings to factors?}

\item{verbose}{Show progress messages?}
}
\value{
A wide form dataframe
}
\description{
Cast a long dataframe into a wide dataframe
}
