% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMeansTrimmed.R
\name{colMeansTrimmed}
\alias{colMeansTrimmed}
\title{Calculate trimmed means by column}
\usage{
colMeansTrimmed(m, trim = 0.25, na.rm = F)
}
\arguments{
\item{m}{A numeric matrix}

\item{trim}{The fraction (0 to 0.5) of observations to be trimmed from each end of `m` before the
mean is computed. Values of trim outside that range are taken as the nearest endpoint}

\item{na.rm}{Should missing values (including NaN) be omitted from the calculations?}
}
\value{
A numeric vector
}
\description{
Calculate trimmed means by column
}
