% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.tests.R
\name{contingencyMatrix}
\alias{contingencyMatrix}
\title{Create multiple contingency matrices}
\usage{
contingencyMatrix(x, y, use.totals = FALSE, show.warnings = TRUE)
}
\arguments{
\item{x}{A logical matrix}

\item{y}{A logical vector indicating the case/control group (i.e. response)}

\item{use.totals}{Should case/control totals be used instead of case.false and ctrl.false?}

\item{show.warnings}{Show warnings?}
}
\value{
A matrix with the columns: case.true, case.false, ctrl.true, ctrl.false. Each row is
thus one contingency matrix
}
\description{
Create multiple contingency matrices
}
