% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effsize.cramerV.R
\name{cramerV}
\alias{cramerV}
\alias{cramerV.default}
\alias{cramerV.matrix}
\alias{cramerV.data.frame}
\title{Calculate Cramer's V}
\usage{
cramerV(x, ...)

\method{cramerV}{default}(case.true, case.false, ctrl.true, ctrl.false, show.sign = T)

\method{cramerV}{matrix}(m)

\method{cramerV}{data.frame}(m)
}
\arguments{
\item{case.true}{Case group responders}

\item{case.false}{Case group non responders}

\item{ctrl.true}{Control group responders}

\item{ctrl.false}{Control group non responders}

\item{show.sign}{Add `-` to values where effect decreases?}
}
\value{
A numeric vector
}
\description{
Calculate effect sizes using Cramer's V values for pairwise categorical variable
comparisons (e.g. comparisons that would be done by a fisher's exact test)
}
