% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.fisher.R
\name{fisherChi2}
\alias{fisherChi2}
\alias{fisherChi2.default}
\alias{fisherChi2.matrix}
\alias{fisherChi2.data.frame}
\title{Fast chi-squared test for 2x2 matrices (similar to a fisher test)}
\usage{
fisherChi2(x, ...)

\method{fisherChi2}{default}(
  case.true,
  case.false,
  ctrl.true,
  ctrl.false,
  correct = TRUE,
  return.statistic = FALSE
)

\method{fisherChi2}{matrix}(m, ...)

\method{fisherChi2}{data.frame}(m, ...)
}
\arguments{
\item{...}{Arguments that can be passed to fisherChi2.default()}

\item{case.true}{Group 'x' responders}

\item{case.false}{Group 'x' non responders}

\item{ctrl.true}{Group 'y' responders}

\item{ctrl.false}{Group 'y' non responders}

\item{correct}{a logical indicating whether to apply continuity correction when computing the
test statistic for 2 by 2 tables: one half is subtracted from all |O - E| differences; however,
the correction will not be bigger than the differences themselves.}

\item{return.statistic}{Return the raw chi-square statistic}

\item{m}{A matrix containing the 4 columns corresponding to a contingency matrix}

\item{verbose}{Show progress?}
}
\value{
A numeric vector of pvalues or chi-squared statistics
}
\description{
Fast chi-squared test for 2x2 matrices (similar to a fisher test)
}
