% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.fisher.R
\name{fisherTest}
\alias{fisherTest}
\alias{fisherTest.default}
\alias{fisherTest.matrix}
\alias{fisherTest.data.frame}
\title{Fast fisher's exact tests}
\usage{
fisherTest(x, ...)

\method{fisherTest}{default}(
  case.true,
  case.false,
  ctrl.true,
  ctrl.false,
  alternative = "two.sided",
  verbose = F
)

\method{fisherTest}{matrix}(m, ...)

\method{fisherTest}{data.frame}(m, ...)
}
\arguments{
\item{...}{Arguments that can be passed to fisherTest.default()}

\item{case.true}{Case group (treated) responders}

\item{case.false}{Case group (treated) non responders}

\item{ctrl.true}{Control group (untreated) responders}

\item{ctrl.false}{Control group (untreated) non responders}

\item{alternative}{Indicates the alternative hypothesis and must be one of "two.sided", "greater"
or "less"}

\item{verbose}{Show progress?}

\item{m}{A matrix containing the 4 columns corresponding to a contingency matrix:}
}
\value{
A numeric vector of pvalues
}
\description{
Fast fisher's exact tests
}
