% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.wilcox.R
\name{wilcoxTest}
\alias{wilcoxTest}
\alias{wilcoxTest.matrix}
\alias{wilcoxTest.data.frame}
\alias{wilcoxTest.default}
\title{Perform multiple wilcox tests}
\usage{
wilcoxTest(x, ...)

\method{wilcoxTest}{matrix}(
  x,
  y,
  alternative = "two.sided",
  mu = 0,
  exact = NA,
  correct = TRUE,
  pvalue.only = TRUE
)

\method{wilcoxTest}{data.frame}(
  x,
  y,
  alternative = "two.sided",
  mu = 0,
  exact = NA,
  correct = TRUE,
  pvalue.only = TRUE
)

\method{wilcoxTest}{default}(x, y, ...)
}
\arguments{
\item{x}{numeric matrix.}

\item{y}{numeric matrix for the second group of observations.}

\item{alternative}{alternative hypothesis to use for each row/column of x.
A single string or a vector with values for each observation.
Values must be one of "two.sided" (default), "greater" or "less".}

\item{mu}{true values of the location shift for the null hypothesis.
A single number or numeric vector with values for each observation.}

\item{exact}{logical or NA (default) indicator whether an exact p-value
should be computed (see Details).
A single value or a logical vector with values for each observation.}

\item{correct}{logical indicator whether continuity correction should be
applied in the cases where p-values are obtained using normal approximation.
A single value or logical vector with values for each observation.}

\item{pvalue.only}{Only return pvalues?}
}
\value{
If pvalue.only==TRUE, a numeric vector of pvalues
Else, a data.frame where each row contains the results of a wilcoxon test
performed on the corresponding row/column of x.
The columns will vary depending on the type of test performed.\cr\cr
They will contain a subset of the following information:\cr
1. obs.x - number of x observations\cr
2. obs.y - number of y observations\cr
3. obs.tot - total number of observations\cr
4. obs.paired - number of paired observations (present in x and y)\cr
5. statistic - Wilcoxon test statistic\cr
6. pvalue - p-value\cr
7. alternative - chosen alternative hypothesis\cr
8. location.null - location shift of the null hypothesis\cr
9. exact - indicates if exact p-value was computed\cr
10. correct - indicates if continuity correction was performed
}
\description{
Perform column-wise wilcox tests. Adapted from \
https://github.com/karoliskoncevicius/matrixTests/
}
