% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyMatrix.R
\name{contingencyMatrix}
\alias{contingencyMatrix}
\title{Create multiple contingency matrices}
\usage{
contingencyMatrix(
  x,
  y,
  use.totals = FALSE,
  avg.func = c("mean", "iqm", "median"),
  show.warnings = TRUE
)
}
\arguments{
\item{x}{A logical or numeric vector/matrix}

\item{y}{A logical vector indicating the case/control group (i.e. response)}

\item{use.totals}{Should case/control totals be used instead of case.false and ctrl.false? Only
applies when `x` is logical}

\item{avg.func}{Can be mean, iqm, or median. When `x` is numeric, the averages in the case and
ctrl groups are calculated. The resulting contigency matrix is then: (i) the number of case
samples above/below the ctrl average, and (ii) the number of ctrl samples above/below the case
average}

\item{show.warnings}{Show warnings?}
}
\value{
A matrix with the columns: case.true, case.false, ctrl.true, ctrl.false. Each row is
thus one contingency matrix
}
\description{
Create multiple contingency matrices
}
